Move logic to the feature label itself
