+++ plugins/svg/src/SvgFeatureRenderer/components/FeatureLabel.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { observer } from 'mobx-react'
-import { measureText } from '@jbrowse/core/util'
+import { isStateTreeNode } from 'mobx-state-tree'
+import { measureText, getViewParams } from '@jbrowse/core/util'
 import { DisplayModel } from './util'
 
 export default observer(
@@ -13,6 +14,10 @@ export default observer(
     fontHeight = 13,
     featureWidth = 0,
     allowedWidthExpansion = 0,
+    feature,
+    viewStart,
+    viewEnd,
+    viewOffsetPx,
     displayModel = {},
   }: {
     text: string
@@ -27,6 +32,28 @@ export default observer(
   }) => {
     const totalWidth = featureWidth + allowedWidthExpansion
     const measuredTextWidth = measureText(text, fontHeight)
+
+    if (isStateTreeNode(displayModel)) {
+      const params = getViewParams(displayModel)
+      viewStart = params.viewStart
+      viewEnd = params.viewEnd
+      viewOffsetPx = params.viewOffsetPx
+    }
+    const viewLeft = reversed ? viewEnd : viewStart
+
+    const fstart = feature.get('start')
+    const fend = feature.get('end')
+    // const [fstart, fend] = reversed ? [end, start] : [start, end]
+    const w = featureWidth
+    if (reversed) {
+      if (fstart < viewLeft + w && viewLeft - w < fend) {
+        x = viewOffsetPx
+      }
+    } else {
+      if (fstart < viewLeft + w && viewLeft - w < fend) {
+        x = viewOffsetPx
+      }
+    }
 
     return (
       <text x={x} y={y + fontHeight} fill={color} fontSize={fontHeight}>
